report, detail report status

+++ VolcomInventory/FormDelManifestDet.vb
@@ -28,12 +28,13 @@
 
     Sub form_load()
         Dim query As String = "
-            SELECT m.id_del_manifest, m.id_comp, m.number, DATE_FORMAT(m.created_date, '%d %M %Y %H:%i:%s') AS created_date, DATE_FORMAT(m.updated_date, '%d %M %Y %H:%i:%s') AS updated_date, ea.employee_name AS created_by, eb.employee_name AS updated_by, m.id_report_status, IFNULL(m.id_report_status, 'Draft') AS report_status
+            SELECT m.id_del_manifest, m.id_comp, m.number, DATE_FORMAT(m.created_date, '%d %M %Y %H:%i:%s') AS created_date, DATE_FORMAT(m.updated_date, '%d %M %Y %H:%i:%s') AS updated_date, ea.employee_name AS created_by, eb.employee_name AS updated_by, m.id_report_status, IFNULL(l.report_status, 'Draft') AS report_status
             FROM tb_del_manifest AS m
             LEFT JOIN tb_m_user AS ua ON m.created_by = ua.id_user
             LEFT JOIN tb_m_employee AS ea ON ua.id_employee = ea.id_employee
             LEFT JOIN tb_m_user AS ub ON m.created_by = ub.id_user
             LEFT JOIN tb_m_employee AS eb ON ub.id_employee = eb.id_employee
+            LEFT JOIN tb_lookup_report_status AS l ON m.id_report_status = l.id_report_status
             WHERE m.id_del_manifest = " + id_del_manifest + "
 
             UNION
@@ -97,6 +98,12 @@
             SBAdd.Enabled = False
             SBRemove.Enabled = False
         End If
+
+        SBAttachement.Enabled = True
+
+        If id_del_manifest = "0" Then
+            SBAttachement.Enabled = False
+        End If
     End Sub
 
     Sub view_3pl()
@@ -109,43 +116,59 @@
         If SLUE3PL.EditValue.ToString = "0" Then
             stopCustom("Please select 3PL.")
         Else
-            Dim query As String = ""
+            Dim continue_save As Boolean = True
 
-            If id_del_manifest = "0" Then
-                query = "INSERT INTO tb_del_manifest (id_comp, created_date, created_by) VALUES (" + SLUE3PL.EditValue.ToString + ", NOW(), " + id_user + "); SELECT LAST_INSERT_ID();"
+            If type = "complete" Or type = "cancel" Then
+                Dim confirm As DialogResult
 
-                id_del_manifest = execute_query(query, 0, True, "", "", "", "")
-            Else
-                'update
-                query = "UPDATE tb_del_manifest SET id_comp = " + SLUE3PL.EditValue.ToString + ", updated_date = NOW(), updated_by = " + id_user + ", id_report_status = " + If(type = "draft", "NULL", If(type = "complete", "6", "5")) + " WHERE id_del_manifest = " + id_del_manifest
+                confirm = DevExpress.XtraEditors.XtraMessageBox.Show("Are you sure want to " + type + " ?", "Warning", MessageBoxButtons.YesNo, MessageBoxIcon.Warning, MessageBoxDefaultButton.Button2)
 
-                execute_non_query(query, True, "", "", "", "")
-
-                'delete
-                query = "DELETE FROM tb_del_manifest_det WHERE id_del_manifest = " + id_del_manifest
-
-                execute_non_query(query, True, "", "", "", "")
+                If confirm = Windows.Forms.DialogResult.Yes Then
+                    continue_save = True
+                Else
+                    continue_save = False
+                End If
             End If
 
-            'detail
-            query = "INSERT INTO tb_del_manifest_det (id_del_manifest, id_wh_awb_det) VALUES "
+            If continue_save Then
+                Dim query As String = ""
 
-            For i = 0 To GVList.RowCount - 1
-                If GVList.IsValidRowHandle(i) Then
-                    query += "(" + id_del_manifest + ", " + GVList.GetRowCellValue(i, "id_wh_awb_det").ToString + "), "
+                If id_del_manifest = "0" Then
+                    query = "INSERT INTO tb_del_manifest (id_comp, created_date, created_by) VALUES (" + SLUE3PL.EditValue.ToString + ", NOW(), " + id_user + "); SELECT LAST_INSERT_ID();"
+
+                    id_del_manifest = execute_query(query, 0, True, "", "", "", "")
+                Else
+                    'update
+                    query = "UPDATE tb_del_manifest SET id_comp = " + SLUE3PL.EditValue.ToString + ", updated_date = NOW(), updated_by = " + id_user + ", id_report_status = " + If(type = "draft", "NULL", If(type = "complete", "6", "5")) + " WHERE id_del_manifest = " + id_del_manifest
+
+                    execute_non_query(query, True, "", "", "", "")
+
+                    'delete
+                    query = "DELETE FROM tb_del_manifest_det WHERE id_del_manifest = " + id_del_manifest
+
+                    execute_non_query(query, True, "", "", "", "")
                 End If
-            Next
 
-            query = query.Substring(0, query.Length - 2)
+                'detail
+                query = "INSERT INTO tb_del_manifest_det (id_del_manifest, id_wh_awb_det) VALUES "
 
-            execute_non_query(query, True, "", "", "", "")
+                For i = 0 To GVList.RowCount - 1
+                    If GVList.IsValidRowHandle(i) Then
+                        query += "(" + id_del_manifest + ", " + GVList.GetRowCellValue(i, "id_wh_awb_det").ToString + "), "
+                    End If
+                Next
 
-            execute_non_query("CALL gen_number(" + id_del_manifest + ", '232')", True, "", "", "", "")
+                query = query.Substring(0, query.Length - 2)
 
-            If type = "draft" Then
-                form_load()
-            Else
-                Close()
+                execute_non_query(query, True, "", "", "", "")
+
+                execute_non_query("CALL gen_number(" + id_del_manifest + ", '232')", True, "", "", "", "")
+
+                If type = "draft" Then
+                    form_load()
+                Else
+                    Close()
+                End If
             End If
         End If
     End Sub
@@ -159,6 +182,40 @@
     End Sub
 
     Private Sub SBPrint_Click(sender As Object, e As EventArgs) Handles SBPrint.Click
+        Dim report As ReportDelManifest = New ReportDelManifest
 
+        report.id_del_manifest = id_del_manifest
+        report.dt = GCList.DataSource
+
+        report.XrLabelNumber.Text = TENumber.Text
+        report.XrLabel3PL.Text = SLUE3PL.Text
+
+        Dim tool As DevExpress.XtraReports.UI.ReportPrintTool = New DevExpress.XtraReports.UI.ReportPrintTool(report)
+
+        tool.ShowPreview()
+    End Sub
+
+    Private Sub SBAttachement_Click(sender As Object, e As EventArgs) Handles SBAttachement.Click
+        Cursor = Cursors.WaitCursor
+
+        Dim id_report_status As String = execute_query("SELECT IFNULL((SELECT id_report_status FROM tb_del_manifest WHERE id_del_manifest = " + id_del_manifest + "), 0) AS id_report_status", 0, True, "", "", "", "")
+
+        FormDocumentUpload.is_no_delete = If(Not id_report_status = "0", "1", "-1")
+        FormDocumentUpload.is_view = If(Not id_report_status = "0", "1", "-1")
+        FormDocumentUpload.id_report = id_del_manifest
+        FormDocumentUpload.report_mark_type = "232"
+
+        FormDocumentUpload.ShowDialog()
+
+        Cursor = Cursors.Default
+    End Sub
+
+    Private Sub FormDelManifestDet_FormClosed(sender As Object, e As FormClosedEventArgs) Handles MyBase.FormClosed
+        Try
+            FormDelManifest.form_load()
+        Catch ex As Exception
+        End Try
+
+        Dispose()
     End Sub
 End Class
